pyproject.toml
[tool.poetry]
name = "rainbow_fatcat"
version = "0.1.0"
description = ""
authors = ["Elton Chou <plscd748@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9"
"discord.py" = "^2.0.1"
eorzeaenv = "^2.2.4"
pydantic = "^1.9.2"
Babel = "^2.10.3"

[tool.poetry.dev-dependencies]
autopep8 = "^1.7.0"
flake8 = "^5.0.4"
mypy = "^0.971"
pre-commit = "^2.20.0"
black = "^22.8.0"

[tool.mypy]
files = ["rainbow_fatcat/**/*.py"]
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
